changeVariable type int to double

+++ 3rdSemesterLabReportPrograms/OOP/05.answer.cpp
@@ -9,11 +9,11 @@ ld volume(double x) { return x * x * x; }
 int main()
 {
     std::cout << "Enter the Length, Breadth,height of the Rectangle: ";
-    int length, breadth, height;
+    double length, breadth, height;
     std::cin >> length >> breadth >> height;
     std::cout << "The Volume of the Rectangle is: " << volume(length, breadth, height) << std::endl;
     std::cout << "Enter the Radius and Height of the Cylinder: ";
-    int radius;
+    double radius;
     std::cin >> radius >> height;
     std::cout << "The Volume of the Cylinder is: " << volume(radius, height) << std::endl;
     std::cout << "Enter the length of a side of The Cube: ";
